refactor: decompose god files and harden npm install scripts
Split three large files into focused sub-modules (all tests pass, no consumer imports changed):
- routes/cos.js (1,464 lines) → 9 route sub-modules (cosStatusRoutes, cosTaskRoutes, cosAgentRoutes, cosReportRoutes, cosLearningRoutes, cosScheduleRoutes, cosJobRoutes, cosTemplateRoutes, cosInsightRoutes)
- routes/scaffold.js (1,667 lines) → 4 template generators (scaffoldVite, scaffoldExpress, scaffoldIOS, scaffoldPortOS)
- client/src/services/api.js (2,016 lines) → 27 domain sub-modules with barrel re-export

Add .npmrc with ignore-scripts=true; install:all now explicitly rebuilds trusted packages (esbuild, node-pty) after install.
