+++ client/src/services/api.js
@@ -1,2016 +1,34 @@
-import toast from '../components/ui/Toast';
-
-const API_BASE = '/api';
-
-// Stable ID for the PortOS baseline app (mirrors server PORTOS_APP_ID)
-export const PORTOS_APP_ID = 'portos-default';
-
-async function request(endpoint, options = {}) {
-  const { silent, ...fetchOptions } = options;
-  const url = `${API_BASE}${endpoint}`;
-  const config = {
-    headers: {
-      'Content-Type': 'application/json',
-      ...fetchOptions.headers
-    },
-    ...fetchOptions
-  };
-
-  const response = await fetch(url, config).catch(() => null);
-  if (!response) {
-    const msg = 'Server unreachable — check your connection and try again';
-    if (!silent) toast.error(msg);
-    throw new Error(msg);
-  }
-
-  if (!response.ok) {
-    const error = await response.json().catch(() => ({ error: 'Request failed' }));
-    const errorMessage = error.error || `HTTP ${response.status}`;
-    if (!silent) {
-      // Platform unavailability is a warning, not an error
-      if (error.code === 'PLATFORM_UNAVAILABLE') {
-        toast(errorMessage, { icon: '⚠️' });
-      } else {
-        toast.error(errorMessage);
-      }
-    }
-    throw new Error(errorMessage);
-  }
-
-  // Handle 204 No Content
-  if (response.status === 204) {
-    return null;
-  }
-
-  return response.json();
-}
-
-// Search
-export const search = (q) => request(`/search?q=${encodeURIComponent(q)}`);
-
-// Apple Health
-export const ingestAppleHealth = (data) => request('/health/ingest', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const getAppleHealthMetrics = (metricName, from, to) => {
-  const params = new URLSearchParams();
-  if (from) params.set('from', from);
-  if (to) params.set('to', to);
-  return request(`/health/metrics/${metricName}/daily?${params}`);
-};
-export const getAppleHealthSummary = (metricName, from, to) => {
-  const params = new URLSearchParams();
-  if (from) params.set('from', from);
-  if (to) params.set('to', to);
-  return request(`/health/metrics/${metricName}?${params}`);
-};
-export const getAvailableHealthMetrics = () => request('/health/metrics/available');
-export const getLatestHealthMetrics = (metricNames) =>
-  request(`/health/metrics/latest?metrics=${metricNames.join(',')}`);
-export const getAppleHealthRange = () => request('/health/range');
-export const getAppleHealthCorrelation = (from, to) => {
-  const params = new URLSearchParams();
-  if (from) params.set('from', from);
-  if (to) params.set('to', to);
-  return request(`/health/correlation?${params}`);
-};
-export const uploadAppleHealthXml = (file) => {
-  const formData = new FormData();
-  formData.append('file', file);
-  // Use fetch directly — the request helper sets Content-Type: application/json
-  // which conflicts with multipart/form-data. Browser sets correct boundary automatically.
-  return fetch(`${API_BASE}/health/import/xml`, {
-    method: 'POST',
-    body: formData,
-  }).then(async res => {
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({ error: res.statusText }));
-      throw new Error(err.error || res.statusText);
-    }
-    return res.json();
-  });
-};
-
-// Alerts
-export const getAlertsSummary = (options) => request('/alerts/summary', options);
-
-// Health
-export const checkHealth = () => request('/system/health');
-export const getSystemHealth = (options) => request('/system/health/details', options);
-
-// Update
-export const getUpdateStatus = () => request('/update/status');
-export const checkForUpdate = () => request('/update/check', { method: 'POST' });
-export const ignoreUpdateVersion = (version) => request('/update/ignore', {
-  method: 'POST',
-  body: JSON.stringify({ version })
-});
-export const clearIgnoredVersions = () => request('/update/ignore', { method: 'DELETE' });
-export const executePortosUpdate = () => request('/update/execute', { method: 'POST' });
-
-// Apps
-export const getApps = () => request('/apps');
-export const getApp = (id) => request(`/apps/${id}`);
-export const createApp = (data) => request('/apps', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateApp = (id, data) => request(`/apps/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteApp = (id) => request(`/apps/${id}`, { method: 'DELETE' });
-
-// App actions
-export const startApp = (id) => request(`/apps/${id}/start`, { method: 'POST' });
-export const stopApp = (id) => request(`/apps/${id}/stop`, { method: 'POST' });
-export const restartApp = (id) => request(`/apps/${id}/restart`, { method: 'POST' });
-
-/**
- * Handle PortOS self-restart: show a loading toast, poll for server recovery, then reload.
- * Call this after restartApp() returns { selfRestart: true }.
- */
-export function handleSelfRestart() {
-  toast.loading('Restarting PortOS...', { id: 'self-restart', duration: Infinity });
-  const poll = async () => {
-    for (let i = 0; i < 30; i++) {
-      await new Promise(r => setTimeout(r, 2000));
-      const ok = await fetch(`${API_BASE}/system/health`).then(() => true).catch(() => false);
-      if (ok) {
-        toast.success('PortOS restarted successfully', { id: 'self-restart' });
-        setTimeout(() => window.location.reload(), 1000);
-        return;
-      }
-    }
-    toast.error('PortOS restart timed out — try reloading manually', { id: 'self-restart' });
-  };
-  poll();
-}
-export const archiveApp = (id) => request(`/apps/${id}/archive`, { method: 'POST' });
-export const unarchiveApp = (id) => request(`/apps/${id}/unarchive`, { method: 'POST' });
-export const openAppInEditor = (id) => request(`/apps/${id}/open-editor`, { method: 'POST' });
-export const openAppInClaude = (id) => request(`/apps/${id}/open-claude`, { method: 'POST' });
-export const openAppFolder = (id) => request(`/apps/${id}/open-folder`, { method: 'POST' });
-export const refreshAppConfig = (id) => request(`/apps/${id}/refresh-config`, { method: 'POST' });
-export const pullAndUpdateApp = (id) => request(`/apps/${id}/update`, { method: 'POST' });
-export const buildApp = (id) => request(`/apps/${id}/build`, { method: 'POST' });
-export const getAppStatus = (id) => request(`/apps/${id}/status`);
-export const getAppTaskTypes = (id) => request(`/apps/${id}/task-types`);
-export const toggleAllAppTaskTypes = (id, enabled) => request(`/apps/${id}/task-types/all`, {
-  method: 'PUT',
-  body: JSON.stringify({ enabled })
-});
-export const updateAppTaskTypeOverride = (id, taskType, { enabled, interval, taskMetadata } = {}) => request(`/apps/${id}/task-types/${taskType}`, {
-  method: 'PUT',
-  body: JSON.stringify({ enabled, interval, taskMetadata })
-});
-export const bulkUpdateAppTaskTypeOverride = (taskType, { enabled }) => request(`/apps/bulk-task-type/${taskType}`, {
-  method: 'PUT',
-  body: JSON.stringify({ enabled })
-});
-export const detectAppIcons = () => request('/apps/detect-icons', { method: 'POST' });
-export const getAppLogs = (id, lines = 100, processName) => {
-  const params = new URLSearchParams({ lines: String(lines) });
-  if (processName) params.set('process', processName);
-  return request(`/apps/${id}/logs?${params}`);
-};
-
-export const getAppDocuments = (id) => request(`/apps/${id}/documents`);
-export const getAppDocument = (id, filename) => request(`/apps/${id}/documents/${filename}`);
-export const saveAppDocument = (id, filename, content, commitMessage) =>
-  request(`/apps/${id}/documents/${filename}`, {
-    method: 'PUT',
-    body: JSON.stringify({ content, ...(commitMessage && { commitMessage }) })
-  });
-export const getAppAgents = (id, limit = 50) => request(`/apps/${id}/agents?limit=${limit}`);
-
-// Ports
-export const scanPorts = () => request('/ports/scan');
-export const checkPorts = (ports) => request('/ports/check', {
-  method: 'POST',
-  body: JSON.stringify({ ports })
-});
-export const allocatePorts = (count = 1) => request('/ports/allocate', {
-  method: 'POST',
-  body: JSON.stringify({ count })
-});
-
-// Detect
-export const detectRepo = (path) => request('/detect/repo', {
-  method: 'POST',
-  body: JSON.stringify({ path })
-});
-
-export const detectPort = (port) => request('/detect/port', {
-  method: 'POST',
-  body: JSON.stringify({ port })
-});
-
-export const detectPm2 = (name) => request('/detect/pm2', {
-  method: 'POST',
-  body: JSON.stringify({ name })
-});
-
-export const detectWithAi = (path, providerId) => request('/detect/ai', {
-  method: 'POST',
-  body: JSON.stringify({ path, providerId })
-});
-
-// Templates & Scaffold
-export const getTemplates = () => request('/scaffold/templates');
-
-export const getDirectories = (path = null) => {
-  const params = path ? `?path=${encodeURIComponent(path)}` : '';
-  return request(`/scaffold/directories${params}`);
-};
-
-export const scaffoldApp = (data) => request('/scaffold', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-
-export const createFromTemplate = (data) => request('/scaffold/templates/create', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-
-// Providers
-export const getProviders = () => request('/providers');
-export const getActiveProvider = () => request('/providers/active');
-export const setActiveProvider = (id) => request('/providers/active', {
-  method: 'PUT',
-  body: JSON.stringify({ id })
-});
-export const getProvider = (id) => request(`/providers/${id}`);
-export const createProvider = (data) => request('/providers', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateProvider = (id, data) => request(`/providers/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteProvider = (id) => request(`/providers/${id}`, { method: 'DELETE' });
-export const getSampleProviders = () => request('/providers/samples');
-export const testProvider = (id) => request(`/providers/${id}/test`, { method: 'POST' });
-export const refreshProviderModels = (id) => request(`/providers/${id}/refresh-models`, { method: 'POST' });
-
-// Provider status (usage limits, availability)
-export const getProviderStatuses = () => request('/providers/status');
-export const getProviderStatus = (id) => request(`/providers/${id}/status`);
-export const recoverProvider = (id) => request(`/providers/${id}/status/recover`, { method: 'POST' });
-
-// Runs
-export const getRuns = (limit = 50, offset = 0, source = 'all') =>
-  request(`/runs?limit=${limit}&offset=${offset}&source=${source}`);
-export const createRun = (data) => request('/runs', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const getRun = (id) => request(`/runs/${id}`);
-export const getRunOutput = (id) => request(`/runs/${id}/output`);
-export const getRunPrompt = (id) => request(`/runs/${id}/prompt`);
-export const stopRun = (id) => request(`/runs/${id}/stop`, { method: 'POST' });
-export const deleteRun = (id) => request(`/runs/${id}`, { method: 'DELETE' });
-export const deleteFailedRuns = () => request('/runs?filter=failed', { method: 'DELETE' });
-
-// History
-export const getHistory = (options = {}) => {
-  const params = new URLSearchParams();
-  if (options.limit) params.set('limit', options.limit);
-  if (options.offset) params.set('offset', options.offset);
-  if (options.action) params.set('action', options.action);
-  if (options.success !== undefined) params.set('success', options.success);
-  return request(`/history?${params}`);
-};
-export const getHistoryStats = () => request('/history/stats');
-export const getHistoryActions = () => request('/history/actions');
-export const clearHistory = (olderThanDays) => request(
-  olderThanDays ? `/history?olderThanDays=${olderThanDays}` : '/history',
-  { method: 'DELETE' }
-);
-export const deleteHistoryEntry = (id) => request(`/history/${id}`, { method: 'DELETE' });
-
-// Commands
-export const executeCommand = (command, workspacePath) => request('/commands/execute', {
-  method: 'POST',
-  body: JSON.stringify({ command, workspacePath })
-});
-export const stopCommand = (id) => request(`/commands/${id}/stop`, { method: 'POST' });
-export const getAllowedCommands = () => request('/commands/allowed');
-export const getProcessesList = () => request('/commands/processes');
-
-// Git
-export const getGitInfo = (path) => request('/git/info', {
-  method: 'POST',
-  body: JSON.stringify({ path })
-});
-export const getGitStatus = (path) => request('/git/status', {
-  method: 'POST',
-  body: JSON.stringify({ path })
-});
-export const getGitDiff = (path, staged = false) => request('/git/diff', {
-  method: 'POST',
-  body: JSON.stringify({ path, staged })
-});
-export const getGitCommits = (path, limit = 10) => request('/git/commits', {
-  method: 'POST',
-  body: JSON.stringify({ path, limit })
-});
-export const stageFiles = (path, files) => request('/git/stage', {
-  method: 'POST',
-  body: JSON.stringify({ path, files })
-});
-export const unstageFiles = (path, files) => request('/git/unstage', {
-  method: 'POST',
-  body: JSON.stringify({ path, files })
-});
-export const createCommit = (path, message) => request('/git/commit', {
-  method: 'POST',
-  body: JSON.stringify({ path, message })
-});
-export const updateBranches = (path) => request('/git/update-branches', {
-  method: 'POST',
-  body: JSON.stringify({ path })
-});
-export const getBranchComparison = (path, base, head) => request('/git/branch-comparison', {
-  method: 'POST',
-  body: JSON.stringify({ path, base, head })
-});
-export const pushBranch = (path, branch) => request('/git/push', {
-  method: 'POST',
-  body: JSON.stringify({ path, branch })
-});
-export const pushAllBranches = (path) => request('/git/push-all', {
-  method: 'POST',
-  body: JSON.stringify({ path })
-});
-export const getBranches = (path) => request('/git/branches', {
-  method: 'POST',
-  body: JSON.stringify({ path })
-});
-export const checkoutBranch = (path, branch) => request('/git/checkout', {
-  method: 'POST',
-  body: JSON.stringify({ path, branch })
-});
-export const pullBranch = (path) => request('/git/pull', {
-  method: 'POST',
-  body: JSON.stringify({ path })
-});
-export const syncBranch = (path, branch) => request('/git/sync', {
-  method: 'POST',
-  body: JSON.stringify({ path, branch })
-});
-export const getRemoteBranches = (path) => request('/git/remote-branches', {
-  method: 'POST',
-  body: JSON.stringify({ path })
-});
-export const deleteBranch = (path, branch, { local = false, remote = false } = {}) =>
-  request('/git/delete-branch', {
-    method: 'POST',
-    body: JSON.stringify({ path, branch, local, remote })
-  });
-export const cleanupMergedBranches = (path) => request('/git/cleanup-merged', {
-  method: 'POST',
-  body: JSON.stringify({ path })
-});
-export const mergeBranch = (path, branch) => request('/git/merge', {
-  method: 'POST',
-  body: JSON.stringify({ path, branch })
-});
-export const checkoutRemoteBranch = (path, branch) => request('/git/checkout-remote', {
-  method: 'POST',
-  body: JSON.stringify({ path, branch })
-});
-
-// Usage
-export const getUsage = () => request('/usage');
-export const getUsageRaw = () => request('/usage/raw');
-export const resetUsage = () => request('/usage', { method: 'DELETE' });
-
-// Screenshots
-export const uploadScreenshot = (base64Data, filename, mimeType) => request('/screenshots', {
-  method: 'POST',
-  body: JSON.stringify({ data: base64Data, filename, mimeType })
-});
-
-// Attachments (generic file uploads for tasks)
-export const uploadAttachment = (base64Data, filename) => request('/attachments', {
-  method: 'POST',
-  body: JSON.stringify({ data: base64Data, filename })
-});
-export const getAttachment = (filename) => request(`/attachments/${encodeURIComponent(filename)}`);
-export const deleteAttachment = (filename) => request(`/attachments/${encodeURIComponent(filename)}`, { method: 'DELETE' });
-export const listAttachments = () => request('/attachments');
-
-// Uploads (general file storage)
-export const uploadFile = (base64Data, filename) => request('/uploads', {
-  method: 'POST',
-  body: JSON.stringify({ data: base64Data, filename })
-});
-export const listUploads = () => request('/uploads');
-export const getUploadUrl = (filename) => `/api/uploads/${encodeURIComponent(filename)}`;
-export const deleteUpload = (filename) => request(`/uploads/${encodeURIComponent(filename)}`, { method: 'DELETE' });
-export const deleteAllUploads = () => request('/uploads?confirm=true', { method: 'DELETE' });
-
-// Running Agents (Process Management)
-export const getRunningAgents = () => request('/agents');
-export const getRunningAgentInfo = (pid) => request(`/agents/${pid}`);
-export const killRunningAgent = (pid) => request(`/agents/${pid}`, { method: 'DELETE' });
-// Legacy aliases
-export const getAgents = getRunningAgents;
-export const getAgentInfo = getRunningAgentInfo;
-export const killAgent = killRunningAgent;
-
-// Agent Personalities
-export const getAgentPersonalities = (userId = null) => {
-  const params = userId ? `?userId=${encodeURIComponent(userId)}` : '';
-  return request(`/agents/personalities${params}`);
-};
-export const getAgentPersonality = (id) => request(`/agents/personalities/${id}`);
-export const createAgentPersonality = (data) => request('/agents/personalities', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateAgentPersonality = (id, data) => request(`/agents/personalities/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteAgentPersonality = (id) => request(`/agents/personalities/${id}`, { method: 'DELETE' });
-export const toggleAgentPersonality = (id, enabled) => request(`/agents/personalities/${id}/toggle`, {
-  method: 'POST',
-  body: JSON.stringify({ enabled })
-});
-export const generateAgentPersonality = (seedData, providerId, model) => request('/agents/personalities/generate', {
-  method: 'POST',
-  body: JSON.stringify({ seed: seedData, providerId, model })
-});
-
-// Platform Accounts
-export const getPlatformAccounts = (agentId = null, platform = null) => {
-  const params = new URLSearchParams();
-  if (agentId) params.set('agentId', agentId);
-  if (platform) params.set('platform', platform);
-  const query = params.toString();
-  return request(`/agents/accounts${query ? `?${query}` : ''}`);
-};
-export const getPlatformAccount = (id) => request(`/agents/accounts/${id}`);
-export const createPlatformAccount = (data) => request('/agents/accounts', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const registerPlatformAccount = (agentId, platform, name, description) => request('/agents/accounts', {
-  method: 'POST',
-  body: JSON.stringify({ agentId, platform, name, description })
-});
-export const deletePlatformAccount = (id) => request(`/agents/accounts/${id}`, { method: 'DELETE' });
-export const testPlatformAccount = (id) => request(`/agents/accounts/${id}/test`, { method: 'POST' });
-export const claimPlatformAccount = (id) => request(`/agents/accounts/${id}/claim`, { method: 'POST' });
-
-// Automation Schedules
-export const getAutomationSchedules = (agentId = null, accountId = null) => {
-  const params = new URLSearchParams();
-  if (agentId) params.set('agentId', agentId);
-  if (accountId) params.set('accountId', accountId);
-  const query = params.toString();
-  return request(`/agents/schedules${query ? `?${query}` : ''}`);
-};
-export const getAutomationSchedule = (id) => request(`/agents/schedules/${id}`);
-export const getScheduleStats = () => request('/agents/schedules/stats');
-export const createAutomationSchedule = (data) => request('/agents/schedules', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateAutomationSchedule = (id, data) => request(`/agents/schedules/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteAutomationSchedule = (id) => request(`/agents/schedules/${id}`, { method: 'DELETE' });
-export const toggleAutomationSchedule = (id, enabled) => request(`/agents/schedules/${id}/toggle`, {
-  method: 'POST',
-  body: JSON.stringify({ enabled })
-});
-export const runAutomationScheduleNow = (id) => request(`/agents/schedules/${id}/run`, { method: 'POST' });
-
-// Agent Tools
-export const generateAgentPost = (agentId, accountId, submolt, providerId, model) => request('/agents/tools/generate-post', {
-  method: 'POST',
-  body: JSON.stringify({ agentId, accountId, submolt, providerId, model })
-});
-export const generateAgentComment = (agentId, accountId, postId, parentId, providerId, model) => request('/agents/tools/generate-comment', {
-  method: 'POST',
-  body: JSON.stringify({ agentId, accountId, postId, parentId, providerId, model })
-});
-export const publishAgentPost = (agentId, accountId, submolt, title, content) => request('/agents/tools/publish-post', {
-  method: 'POST',
-  body: JSON.stringify({ agentId, accountId, submolt, title, content })
-});
-export const publishAgentComment = (agentId, accountId, postId, content, parentId) => request('/agents/tools/publish-comment', {
-  method: 'POST',
-  body: JSON.stringify({ agentId, accountId, postId, content, parentId })
-});
-export const engageAgent = (agentId, accountId, maxComments, maxVotes) => request('/agents/tools/engage', {
-  method: 'POST',
-  body: JSON.stringify({ agentId, accountId, maxComments, maxVotes })
-});
-export const getAgentFeed = (accountId, sort, limit) => {
-  const params = new URLSearchParams({ accountId });
-  if (sort) params.set('sort', sort);
-  if (limit) params.set('limit', limit);
-  return request(`/agents/tools/feed?${params}`);
-};
-export const getAgentRelevantPosts = (agentId, accountId, maxResults) => {
-  const params = new URLSearchParams({ agentId, accountId });
-  if (maxResults) params.set('maxResults', maxResults);
-  return request(`/agents/tools/relevant-posts?${params}`);
-};
-export const getAgentSubmolts = (accountId) => request(`/agents/tools/submolts?accountId=${accountId}`);
-export const getAgentPost = (accountId, postId) => request(`/agents/tools/post/${postId}?accountId=${accountId}`);
-export const getAgentRateLimits = (accountId) => request(`/agents/tools/rate-limits?accountId=${accountId}`);
-export const getAgentPublished = (agentId, accountId, days = 7) =>
-  request(`/agents/tools/published?agentId=${agentId}&accountId=${accountId}&days=${days}`);
-export const checkAgentPosts = (agentId, accountId, days, maxReplies, maxUpvotes) =>
-  request('/agents/tools/check-posts', {
-    method: 'POST',
-    body: JSON.stringify({ agentId, accountId, days, maxReplies, maxUpvotes })
-  });
-
-// Moltworld Tools
-export const moltworldJoin = (accountId, x, y, thinking, say, sayTo, agentId) =>
-  request('/agents/tools/moltworld/join', {
-    method: 'POST',
-    body: JSON.stringify({ accountId, agentId, x, y, thinking, say, sayTo })
-  });
-export const moltworldBuild = (accountId, agentId, x, y, z, type, action) =>
-  request('/agents/tools/moltworld/build', {
-    method: 'POST',
-    body: JSON.stringify({ accountId, agentId, x, y, z, type, action })
-  });
-export const moltworldExplore = (accountId, agentId, x, y, thinking) =>
-  request('/agents/tools/moltworld/explore', {
-    method: 'POST',
-    body: JSON.stringify({ accountId, agentId, x, y, thinking })
-  });
-export const moltworldStatus = (accountId) =>
-  request(`/agents/tools/moltworld/status?accountId=${accountId}`);
-export const moltworldBalance = (accountId) =>
-  request(`/agents/tools/moltworld/balance?accountId=${accountId}`);
-export const moltworldRateLimits = (accountId) =>
-  request(`/agents/tools/moltworld/rate-limits?accountId=${accountId}`);
-export const moltworldThink = (accountId, thought, agentId) =>
-  request('/agents/tools/moltworld/think', {
-    method: 'POST',
-    body: JSON.stringify({ accountId, agentId, thought })
-  });
-export const moltworldSay = (accountId, message, sayTo, agentId) =>
-  request('/agents/tools/moltworld/say', {
-    method: 'POST',
-    body: JSON.stringify({ accountId, agentId, message, ...(sayTo ? { sayTo } : {}) })
-  });
-
-// Moltworld Action Queue
-export const moltworldGetQueue = (agentId) =>
-  request(`/agents/tools/moltworld/queue/${agentId}`);
-export const moltworldAddToQueue = (agentId, actionType, params, scheduledFor) =>
-  request('/agents/tools/moltworld/queue', {
-    method: 'POST',
-    body: JSON.stringify({ agentId, actionType, params, scheduledFor })
-  });
-export const moltworldRemoveFromQueue = (id) =>
-  request(`/agents/tools/moltworld/queue/${id}`, { method: 'DELETE' });
-
-// Moltworld WebSocket Relay
-export const moltworldWsConnect = (accountId) =>
-  request('/agents/tools/moltworld/ws/connect', {
-    method: 'POST',
-    body: JSON.stringify({ accountId })
-  });
-export const moltworldWsDisconnect = () =>
-  request('/agents/tools/moltworld/ws/disconnect', { method: 'POST' });
-export const moltworldWsStatus = () =>
-  request('/agents/tools/moltworld/ws/status');
-export const moltworldWsMove = (x, y, thought) =>
-  request('/agents/tools/moltworld/ws/move', {
-    method: 'POST',
-    body: JSON.stringify({ x, y, ...(thought ? { thought } : {}) })
-  });
-export const moltworldWsThink = (thought) =>
-  request('/agents/tools/moltworld/ws/think', {
-    method: 'POST',
-    body: JSON.stringify({ thought })
-  });
-export const moltworldWsNearby = (radius) =>
-  request('/agents/tools/moltworld/ws/nearby', {
-    method: 'POST',
-    body: JSON.stringify({ ...(radius ? { radius } : {}) })
-  });
-export const moltworldWsInteract = (to, payload) =>
-  request('/agents/tools/moltworld/ws/interact', {
-    method: 'POST',
-    body: JSON.stringify({ to, payload })
-  });
-
-// Agent Drafts
-export const getAgentDrafts = (agentId) => request(`/agents/tools/drafts?agentId=${agentId}`);
-export const createAgentDraft = (data) => request('/agents/tools/drafts', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateAgentDraft = (agentId, draftId, data) => request(`/agents/tools/drafts/${draftId}?agentId=${agentId}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteAgentDraft = (agentId, draftId) => request(`/agents/tools/drafts/${draftId}?agentId=${agentId}`, {
-  method: 'DELETE'
-});
-
-// Agent Activity
-export const getAgentActivities = (limit = 50, agentIds = null, action = null) => {
-  const params = new URLSearchParams();
-  params.set('limit', limit);
-  if (agentIds) params.set('agentIds', agentIds.join(','));
-  if (action) params.set('action', action);
-  return request(`/agents/activity?${params}`);
-};
-export const getAgentActivityTimeline = (limit = 50, agentIds = null, before = null) => {
-  const params = new URLSearchParams();
-  params.set('limit', limit);
-  if (agentIds) params.set('agentIds', agentIds.join(','));
-  if (before) params.set('before', before);
-  return request(`/agents/activity/timeline?${params}`);
-};
-export const getAgentActivityByAgent = (agentId, options = {}) => {
-  const params = new URLSearchParams();
-  if (options.date) params.set('date', options.date);
-  if (options.limit) params.set('limit', options.limit);
-  if (options.offset) params.set('offset', options.offset);
-  if (options.action) params.set('action', options.action);
-  return request(`/agents/activity/agent/${agentId}?${params}`);
-};
-export const getAgentActivityStats = (agentId, days = 7) =>
-  request(`/agents/activity/agent/${agentId}/stats?days=${days}`);
-
-// Chief of Staff
-export const getCosStatus = () => request('/cos');
-export const startCos = () => request('/cos/start', { method: 'POST' });
-export const stopCos = () => request('/cos/stop', { method: 'POST' });
-export const pauseCos = (reason) => request('/cos/pause', {
-  method: 'POST',
-  body: JSON.stringify({ reason })
-});
-export const resumeCos = () => request('/cos/resume', { method: 'POST' });
-export const getCosConfig = () => request('/cos/config');
-export const updateCosConfig = (config) => request('/cos/config', {
-  method: 'PUT',
-  body: JSON.stringify(config)
-});
-export const getCosTasks = () => request('/cos/tasks');
-export const addCosTask = (task) => request('/cos/tasks', {
-  method: 'POST',
-  body: JSON.stringify(task)
-});
-export const enhanceCosTaskPrompt = (data) => request('/cos/tasks/enhance', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateCosTask = (id, updates) => request(`/cos/tasks/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(updates)
-});
-export const deleteCosTask = (id, taskType = 'user') => request(`/cos/tasks/${id}?type=${taskType}`, { method: 'DELETE' });
-export const reorderCosTasks = (taskIds) => request('/cos/tasks/reorder', {
-  method: 'POST',
-  body: JSON.stringify({ taskIds })
-});
-export const approveCosTask = (id) => request(`/cos/tasks/${id}/approve`, { method: 'POST' });
-export const forceCosEvaluate = () => request('/cos/evaluate', { method: 'POST' });
-export const forceSpawnTask = (taskId) => request(`/cos/tasks/${taskId}/spawn`, { method: 'POST' });
-export const getCosHealth = () => request('/cos/health');
-export const forceHealthCheck = () => request('/cos/health/check', { method: 'POST' });
-export const getCosAgents = () => request('/cos/agents');
-export const getCosAgentDates = () => request('/cos/agents/history');
-export const getCosAgentsByDate = (date) => request(`/cos/agents/history/${date}`);
-export const getCosAgent = (id) => request(`/cos/agents/${id}`);
-export const terminateCosAgent = (id) => request(`/cos/agents/${id}/terminate`, { method: 'POST' });
-export const killCosAgent = (id) => request(`/cos/agents/${id}/kill`, { method: 'POST' });
-export const getCosAgentStats = (id) => request(`/cos/agents/${id}/stats`);
-export const deleteCosAgent = (id) => request(`/cos/agents/${id}`, { method: 'DELETE' });
-export const clearCompletedCosAgents = () => request('/cos/agents/completed', { method: 'DELETE' });
-export const submitCosAgentFeedback = (id, feedback) => request(`/cos/agents/${id}/feedback`, {
-  method: 'POST',
-  body: JSON.stringify(feedback)
-});
-export const sendCosAgentBtw = (id, message) => request(`/cos/agents/${id}/btw`, {
-  method: 'POST',
-  body: JSON.stringify({ message })
-});
-export const getCosFeedbackStats = () => request('/cos/feedback/stats');
-export const getCosReports = () => request('/cos/reports');
-export const getCosTodayReport = () => request('/cos/reports/today');
-export const getCosReport = (date) => request(`/cos/reports/${date}`);
-
-// CoS Briefings
-export const getCosBriefings = () => request('/cos/briefings');
-export const getCosLatestBriefing = () => request('/cos/briefings/latest');
-export const getCosBriefing = (date) => request(`/cos/briefings/${date}`);
-
-// CoS Activity
-export const getCosTodayActivity = () => request('/cos/activity/today');
-
-// CoS Learning
-export const getCosLearning = () => request('/cos/learning');
-export const getCosLearningDurations = () => request('/cos/learning/durations');
-export const getCosLearningSkipped = () => request('/cos/learning/skipped');
-export const getCosLearningPerformance = () => request('/cos/learning/performance');
-export const getCosLearningRouting = () => request('/cos/learning/routing');
-export const getCosLearningSummary = (options) => request('/cos/learning/summary', options);
-export const backfillCosLearning = () => request('/cos/learning/backfill', { method: 'POST' });
-export const resetCosTaskTypeLearning = (taskType) => request(`/cos/learning/reset/${encodeURIComponent(taskType)}`, { method: 'POST' });
-
-// CoS Quick Task Templates
-export const getCosTaskTemplates = () => request('/cos/templates');
-export const getCosPopularTemplates = (limit = 5) => request(`/cos/templates/popular?limit=${limit}`);
-export const getCosTemplateCategories = () => request('/cos/templates/categories');
-export const createCosTaskTemplate = (data) => request('/cos/templates', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const createCosTemplateFromTask = (task, templateName) => request('/cos/templates/from-task', {
-  method: 'POST',
-  body: JSON.stringify({ task, templateName })
-});
-export const useCosTaskTemplate = (id) => request(`/cos/templates/${id}/use`, { method: 'POST' });
-export const updateCosTaskTemplate = (id, data) => request(`/cos/templates/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteCosTaskTemplate = (id) => request(`/cos/templates/${id}`, { method: 'DELETE' });
-
-// Weekly Digest
-export const getCosWeeklyDigest = (weekId = null) => {
-  if (weekId) return request(`/cos/digest/${weekId}`);
-  return request('/cos/digest');
-};
-export const listCosWeeklyDigests = () => request('/cos/digest/list');
-export const getCosWeekProgress = () => request('/cos/digest/progress');
-export const getCosDigestText = async () => {
-  const response = await fetch('/api/cos/digest/text');
-  if (!response.ok) throw new Error(`HTTP ${response.status}`);
-  return response.text();
-};
-export const generateCosDigest = (weekId = null) => request('/cos/digest/generate', {
-  method: 'POST',
-  body: JSON.stringify({ weekId })
-});
-export const compareCosWeeks = (week1, week2) => request(`/cos/digest/compare?week1=${week1}&week2=${week2}`);
-
-// Productivity & Streaks
-export const getCosProductivity = () => request('/cos/productivity');
-export const getCosProductivitySummary = () => request('/cos/productivity/summary');
-export const recalculateCosProductivity = () => request('/cos/productivity/recalculate', { method: 'POST' });
-export const getCosProductivityTrends = (days = 30) => request(`/cos/productivity/trends?days=${days}`);
-export const getCosActivityCalendar = (weeks = 12, options) => request(`/cos/productivity/calendar?weeks=${weeks}`, options);
-export const getCosQuickSummary = (options) => request('/cos/quick-summary', options);
-export const getCosRecentTasks = (limit = 10, options) => request(`/cos/recent-tasks?limit=${limit}`, options);
-export const getCosActionableInsights = () => request('/cos/actionable-insights');
-export const getCosGoalProgress = () => request('/cos/goal-progress');
-export const getCosGoalProgressSummary = (options) => request('/cos/goal-progress/summary', options);
-
-// Decision Log
-export const getCosDecisions = (limit = 20, type = null) => {
-  const params = new URLSearchParams({ limit: limit.toString() });
-  if (type) params.append('type', type);
-  return request(`/cos/decisions?${params}`);
-};
-export const getCosDecisionSummary = (options) => request('/cos/decisions/summary', options);
-export const getCosDecisionPatterns = () => request('/cos/decisions/patterns');
-
-// Task Schedule (Configurable Intervals)
-export const getCosUpcomingTasks = (limit = 10) => request(`/cos/upcoming?limit=${limit}`);
-export const getCosSchedule = () => request('/cos/schedule');
-export const getCosScheduleIntervalTypes = () => request('/cos/schedule/interval-types');
-export const getCosScheduleDueTasks = () => request('/cos/schedule/due');
-export const getCosScheduleDueAppTasks = (appId) => request(`/cos/schedule/due/${appId}`);
-// Unified task interval update
-export const updateCosTaskInterval = (taskType, settings) => request(`/cos/schedule/task/${taskType}`, {
-  method: 'PUT',
-  body: JSON.stringify(settings)
-});
-// Deprecated aliases — delegate to unified endpoint
-export const updateCosSelfImprovementInterval = (taskType, settings) => updateCosTaskInterval(taskType, settings);
-export const updateCosAppImprovementInterval = (taskType, settings) => updateCosTaskInterval(taskType, settings);
-
-export const triggerCosOnDemandTask = (taskType, appId = null) => request('/cos/schedule/trigger', {
-  method: 'POST',
-  body: JSON.stringify({ taskType, appId })
-});
-export const getCosOnDemandRequests = () => request('/cos/schedule/on-demand');
-export const resetCosTaskHistory = (taskType, appId = null) => request('/cos/schedule/reset', {
-  method: 'POST',
-  body: JSON.stringify({ taskType, appId })
-});
-export const getCosScheduleTemplates = () => request('/cos/schedule/templates');
-export const addCosScheduleTemplate = (template) => request('/cos/schedule/templates', {
-  method: 'POST',
-  body: JSON.stringify(template)
-});
-export const deleteCosScheduleTemplate = (templateId) => request(`/cos/schedule/templates/${templateId}`, { method: 'DELETE' });
-
-// Autonomous Jobs
-export const getCosJobs = () => request('/cos/jobs');
-export const getCosJobsDue = () => request('/cos/jobs/due');
-export const getCosJobIntervals = () => request('/cos/jobs/intervals');
-export const getCosJobAllowedCommands = () => request('/cos/jobs/allowed-commands');
-export const getCosJob = (id) => request(`/cos/jobs/${id}`);
-export const createCosJob = (data) => request('/cos/jobs', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateCosJob = (id, data) => request(`/cos/jobs/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const toggleCosJob = (id) => request(`/cos/jobs/${id}/toggle`, { method: 'POST' });
-export const triggerCosJob = (id) => request(`/cos/jobs/${id}/trigger`, { method: 'POST' });
-export const deleteCosJob = (id) => request(`/cos/jobs/${id}`, { method: 'DELETE' });
-
-// Memory
-export const getMemories = (options = {}) => {
-  const params = new URLSearchParams();
-  if (options.types) params.set('types', options.types.join(','));
-  if (options.categories) params.set('categories', options.categories.join(','));
-  if (options.tags) params.set('tags', options.tags.join(','));
-  if (options.status) params.set('status', options.status);
-  if (options.appId) params.set('appId', options.appId);
-  if (options.limit) params.set('limit', options.limit);
-  if (options.offset) params.set('offset', options.offset);
-  if (options.sortBy) params.set('sortBy', options.sortBy);
-  if (options.sortOrder) params.set('sortOrder', options.sortOrder);
-  return request(`/memory?${params}`);
-};
-export const getMemory = (id) => request(`/memory/${id}`);
-export const createMemory = (data) => request('/memory', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateMemory = (id, data) => request(`/memory/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteMemory = (id, hard = false) => request(`/memory/${id}?hard=${hard}`, { method: 'DELETE' });
-export const searchMemories = (query, options = {}) => request('/memory/search', {
-  method: 'POST',
-  body: JSON.stringify({ query, ...options })
-});
-export const getMemoryCategories = () => request('/memory/categories');
-export const getMemoryTags = () => request('/memory/tags');
-export const getMemoryTimeline = (options = {}) => {
-  const params = new URLSearchParams();
-  if (options.startDate) params.set('startDate', options.startDate);
-  if (options.endDate) params.set('endDate', options.endDate);
-  if (options.types) params.set('types', options.types.join(','));
-  if (options.limit) params.set('limit', options.limit);
-  return request(`/memory/timeline?${params}`);
-};
-export const getMemoryGraph = () => request('/memory/graph');
-export const getMemoryStats = () => request('/memory/stats');
-export const getRelatedMemories = (id, limit = 10) => request(`/memory/${id}/related?limit=${limit}`);
-export const linkMemories = (sourceId, targetId) => request('/memory/link', {
-  method: 'POST',
-  body: JSON.stringify({ sourceId, targetId })
-});
-export const consolidateMemories = (options = {}) => request('/memory/consolidate', {
-  method: 'POST',
-  body: JSON.stringify(options)
-});
-export const getEmbeddingStatus = () => request('/memory/embeddings/status');
-export const getMemoryBackendStatus = () => request('/memory/backend/status', { silent: true });
-export const approveMemory = (id) => request(`/memory/${id}/approve`, { method: 'POST' });
-export const rejectMemory = (id) => request(`/memory/${id}/reject`, { method: 'POST' });
-
-// Notifications
-export const getNotifications = (options = {}) => {
-  const params = new URLSearchParams();
-  if (options.type) params.set('type', options.type);
-  if (options.unreadOnly) params.set('unreadOnly', 'true');
-  if (options.limit) params.set('limit', options.limit);
-  return request(`/notifications?${params}`);
-};
-export const getNotificationCount = () => request('/notifications/count');
-export const getNotificationCounts = () => request('/notifications/counts');
-export const markNotificationRead = (id) => request(`/notifications/${id}/read`, { method: 'POST' });
-export const markAllNotificationsRead = () => request('/notifications/read-all', { method: 'POST' });
-export const deleteNotification = (id) => request(`/notifications/${id}`, { method: 'DELETE' });
-export const clearNotifications = () => request('/notifications', { method: 'DELETE' });
-
-// PM2 Standardization
-export const analyzeStandardization = (repoPath, providerId) => request('/standardize/analyze', {
-  method: 'POST',
-  body: JSON.stringify({ repoPath, providerId })
-});
-export const analyzeStandardizationByApp = (appId, providerId) => request('/standardize/analyze', {
-  method: 'POST',
-  body: JSON.stringify({ appId, providerId })
-});
-export const applyStandardization = (repoPath, plan) => request('/standardize/apply', {
-  method: 'POST',
-  body: JSON.stringify({ repoPath, plan })
-});
-export const applyStandardizationByApp = (appId, plan) => request('/standardize/apply', {
-  method: 'POST',
-  body: JSON.stringify({ appId, plan })
-});
-export const getStandardizeTemplate = () => request('/standardize/template');
-export const createGitBackup = (repoPath) => request('/standardize/backup', {
-  method: 'POST',
-  body: JSON.stringify({ repoPath })
-});
-
-// Feeds - RSS/Atom Feed Ingestion
-export const getFeeds = () => request('/feeds');
-export const getFeedStats = () => request('/feeds/stats');
-export const getFeedItems = ({ feedId, unreadOnly } = {}) => {
-  const params = new URLSearchParams();
-  if (feedId) params.set('feedId', feedId);
-  if (unreadOnly) params.set('unreadOnly', 'true');
-  return request(`/feeds/items?${params}`);
-};
-export const addFeed = (url) => request('/feeds', {
-  method: 'POST',
-  body: JSON.stringify({ url })
-});
-export const removeFeed = (id) => request(`/feeds/${id}`, { method: 'DELETE' });
-export const refreshFeed = (id) => request(`/feeds/${id}/refresh`, { method: 'POST' });
-export const refreshAllFeeds = () => request('/feeds/refresh-all', { method: 'POST' });
-export const markFeedItemRead = (id) => request(`/feeds/items/${id}/read`, { method: 'POST' });
-export const markAllFeedItemsRead = (feedId) => {
-  const params = feedId ? `?feedId=${feedId}` : '';
-  return request(`/feeds/items/read-all${params}`, { method: 'POST' });
-};
-
-// Brain - Second Brain Feature
-export const getBrainSummary = () => request('/brain/summary');
-export const getBrainSettings = () => request('/brain/settings');
-export const updateBrainSettings = (settings) => request('/brain/settings', {
-  method: 'PUT',
-  body: JSON.stringify(settings)
-});
-
-// Brain - Capture & Inbox
-export const captureBrainThought = (text, providerOverride, modelOverride) => request('/brain/capture', {
-  method: 'POST',
-  body: JSON.stringify({ text, providerOverride, modelOverride })
-});
-export const getBrainInbox = (options = {}) => {
-  const params = new URLSearchParams();
-  if (options.status) params.set('status', options.status);
-  if (options.limit) params.set('limit', options.limit);
-  if (options.offset) params.set('offset', options.offset);
-  return request(`/brain/inbox?${params}`);
-};
-export const getBrainInboxEntry = (id) => request(`/brain/inbox/${id}`);
-export const resolveBrainReview = (inboxLogId, destination, editedExtracted) => request('/brain/review/resolve', {
-  method: 'POST',
-  body: JSON.stringify({ inboxLogId, destination, editedExtracted })
-});
-export const fixBrainClassification = (inboxLogId, newDestination, updatedFields, note) => request('/brain/fix', {
-  method: 'POST',
-  body: JSON.stringify({ inboxLogId, newDestination, updatedFields, note })
-});
-export const retryBrainClassification = (id, providerOverride, modelOverride) => request(`/brain/inbox/${id}/retry`, {
-  method: 'POST',
-  body: JSON.stringify({ providerOverride, modelOverride })
-});
-export const updateBrainInboxEntry = (id, capturedText) => request(`/brain/inbox/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify({ capturedText })
-});
-export const deleteBrainInboxEntry = (id) => request(`/brain/inbox/${id}`, { method: 'DELETE' });
-export const markBrainInboxDone = (id) => request(`/brain/inbox/${id}/done`, { method: 'POST' });
-
-// Brain - People
-export const getBrainPeople = () => request('/brain/people');
-export const getBrainPerson = (id) => request(`/brain/people/${id}`);
-export const createBrainPerson = (data) => request('/brain/people', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateBrainPerson = (id, data) => request(`/brain/people/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteBrainPerson = (id) => request(`/brain/people/${id}`, { method: 'DELETE' });
-
-// Brain - Projects
-export const getBrainProjects = (filters) => {
-  const params = new URLSearchParams();
-  if (filters?.status) params.set('status', filters.status);
-  return request(`/brain/projects?${params}`);
-};
-export const getBrainProject = (id) => request(`/brain/projects/${id}`);
-export const createBrainProject = (data) => request('/brain/projects', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateBrainProject = (id, data) => request(`/brain/projects/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteBrainProject = (id) => request(`/brain/projects/${id}`, { method: 'DELETE' });
-
-// Brain - Ideas
-export const getBrainIdeas = (filters) => {
-  const params = new URLSearchParams();
-  if (filters?.status) params.set('status', filters.status);
-  return request(`/brain/ideas?${params}`);
-};
-export const getBrainIdea = (id) => request(`/brain/ideas/${id}`);
-export const createBrainIdea = (data) => request('/brain/ideas', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateBrainIdea = (id, data) => request(`/brain/ideas/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteBrainIdea = (id) => request(`/brain/ideas/${id}`, { method: 'DELETE' });
-
-// Brain - Admin
-export const getBrainAdmin = (filters) => {
-  const params = new URLSearchParams();
-  if (filters?.status) params.set('status', filters.status);
-  return request(`/brain/admin?${params}`);
-};
-export const getBrainAdminItem = (id) => request(`/brain/admin/${id}`);
-export const createBrainAdminItem = (data) => request('/brain/admin', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateBrainAdminItem = (id, data) => request(`/brain/admin/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteBrainAdminItem = (id) => request(`/brain/admin/${id}`, { method: 'DELETE' });
-
-// Brain - Memories
-export const getBrainMemories = () => request('/brain/memories');
-export const getBrainMemory = (id) => request(`/brain/memories/${id}`);
-export const createBrainMemory = (data) => request('/brain/memories', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateBrainMemory = (id, data) => request(`/brain/memories/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteBrainMemory = (id) => request(`/brain/memories/${id}`, { method: 'DELETE' });
-
-// Brain - Digests & Reviews
-export const getBrainLatestDigest = () => request('/brain/digest/latest');
-export const getBrainDigests = (limit = 10) => request(`/brain/digests?limit=${limit}`);
-export const runBrainDigest = (providerOverride, modelOverride) => request('/brain/digest/run', {
-  method: 'POST',
-  body: JSON.stringify({ providerOverride, modelOverride })
-});
-export const getBrainLatestReview = () => request('/brain/review/latest');
-export const getBrainReviews = (limit = 10) => request(`/brain/reviews?limit=${limit}`);
-export const runBrainReview = (providerOverride, modelOverride) => request('/brain/review/run', {
-  method: 'POST',
-  body: JSON.stringify({ providerOverride, modelOverride })
-});
-
-// Brain - Links
-export const getBrainLinks = (options = {}) => {
-  const params = new URLSearchParams();
-  if (options.linkType) params.set('linkType', options.linkType);
-  if (options.isGitHubRepo !== undefined) params.set('isGitHubRepo', options.isGitHubRepo);
-  if (options.limit) params.set('limit', options.limit);
-  if (options.offset) params.set('offset', options.offset);
-  return request(`/brain/links?${params}`);
-};
-export const getBrainLink = (id) => request(`/brain/links/${id}`);
-export const createBrainLink = (data) => request('/brain/links', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateBrainLink = (id, data) => request(`/brain/links/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteBrainLink = (id) => request(`/brain/links/${id}`, { method: 'DELETE' });
-export const cloneBrainLink = (id) => request(`/brain/links/${id}/clone`, { method: 'POST' });
-export const pullBrainLink = (id) => request(`/brain/links/${id}/pull`, { method: 'POST' });
-export const openBrainLinkFolder = (id) => request(`/brain/links/${id}/open-folder`, { method: 'POST' });
-
-// Brain - Graph
-export const getBrainGraph = () => request('/brain/graph');
-
-// Brain - Bridge Sync (brain data to CoS memory system)
-export const syncBrainData = () => request('/brain/bridge-sync', { method: 'POST' });
-
-// Media - Server media devices
-export const getMediaDevices = () => request('/media/devices');
-export const getMediaStatus = () => request('/media/status');
-export const startMediaStream = (videoDeviceId, audioDeviceId, video = true, audio = true) => request('/media/start', {
-  method: 'POST',
-  body: JSON.stringify({ videoDeviceId, audioDeviceId, video, audio })
-});
-export const stopMediaStream = () => request('/media/stop', { method: 'POST' });
-
-// Digital Twin - Status & Summary
-export const getDigitalTwinStatus = () => request('/digital-twin');
-export const getSoulStatus = getDigitalTwinStatus; // Alias for backwards compatibility
-
-// Digital Twin - Documents
-export const getDigitalTwinDocuments = () => request('/digital-twin/documents');
-export const getSoulDocuments = getDigitalTwinDocuments;
-export const getDigitalTwinDocument = (id) => request(`/digital-twin/documents/${id}`);
-export const getSoulDocument = getDigitalTwinDocument;
-export const createDigitalTwinDocument = (data) => request('/digital-twin/documents', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const createSoulDocument = createDigitalTwinDocument;
-export const updateDigitalTwinDocument = (id, data) => request(`/digital-twin/documents/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const updateSoulDocument = updateDigitalTwinDocument;
-export const deleteDigitalTwinDocument = (id) => request(`/digital-twin/documents/${id}`, { method: 'DELETE' });
-export const deleteSoulDocument = deleteDigitalTwinDocument;
-
-// Digital Twin - Testing
-export const getDigitalTwinTests = () => request('/digital-twin/tests');
-export const getSoulTests = getDigitalTwinTests;
-export const runDigitalTwinTests = (providerId, model, testIds = null) => request('/digital-twin/tests/run', {
-  method: 'POST',
-  body: JSON.stringify({ providerId, model, testIds })
-});
-export const runSoulTests = runDigitalTwinTests;
-export const runDigitalTwinMultiTests = (providers, testIds = null) => request('/digital-twin/tests/run-multi', {
-  method: 'POST',
-  body: JSON.stringify({ providers, testIds })
-});
-export const runSoulMultiTests = runDigitalTwinMultiTests;
-export const getDigitalTwinTestHistory = (limit = 10) => request(`/digital-twin/tests/history?limit=${limit}`);
-export const getSoulTestHistory = getDigitalTwinTestHistory;
-
-// Digital Twin - Enrichment
-export const getDigitalTwinEnrichCategories = () => request('/digital-twin/enrich/categories');
-export const getSoulEnrichCategories = getDigitalTwinEnrichCategories;
-export const getDigitalTwinEnrichProgress = () => request('/digital-twin/enrich/progress');
-export const getSoulEnrichProgress = getDigitalTwinEnrichProgress;
-export const getDigitalTwinEnrichQuestion = (category, providerOverride, modelOverride, skipIndices) => request('/digital-twin/enrich/question', {
-  method: 'POST',
-  body: JSON.stringify({ category, providerOverride, modelOverride, ...(skipIndices?.length ? { skipIndices } : {}) })
-});
-export const getSoulEnrichQuestion = getDigitalTwinEnrichQuestion;
-export const submitDigitalTwinEnrichAnswer = (data) => request('/digital-twin/enrich/answer', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const submitSoulEnrichAnswer = submitDigitalTwinEnrichAnswer;
-
-// Digital Twin - Export
-export const getDigitalTwinExportFormats = () => request('/digital-twin/export/formats');
-export const getSoulExportFormats = getDigitalTwinExportFormats;
-export const exportDigitalTwin = (format, documentIds = null, includeDisabled = false) => request('/digital-twin/export', {
-  method: 'POST',
-  body: JSON.stringify({ format, documentIds, includeDisabled })
-});
-export const exportSoul = exportDigitalTwin;
-
-// Digital Twin - Settings
-export const getDigitalTwinSettings = () => request('/digital-twin/settings');
-export const getSoulSettings = getDigitalTwinSettings;
-export const updateDigitalTwinSettings = (settings) => request('/digital-twin/settings', {
-  method: 'PUT',
-  body: JSON.stringify(settings)
-});
-export const updateSoulSettings = updateDigitalTwinSettings;
-
-// Digital Twin - Validation & Analysis
-export const getDigitalTwinCompleteness = () => request('/digital-twin/validate/completeness');
-export const getSoulCompleteness = getDigitalTwinCompleteness;
-export const detectDigitalTwinContradictions = (providerId, model) => request('/digital-twin/validate/contradictions', {
-  method: 'POST',
-  body: JSON.stringify({ providerId, model })
-});
-export const detectSoulContradictions = detectDigitalTwinContradictions;
-export const generateDigitalTwinTests = (providerId, model) => request('/digital-twin/tests/generate', {
-  method: 'POST',
-  body: JSON.stringify({ providerId, model })
-});
-export const generateSoulTests = generateDigitalTwinTests;
-export const analyzeWritingSamples = (samples, providerId, model) => request('/digital-twin/analyze-writing', {
-  method: 'POST',
-  body: JSON.stringify({ samples, providerId, model })
-});
-
-// Digital Twin - List-based Enrichment
-export const analyzeEnrichmentList = (category, items, providerId, model) => request('/digital-twin/enrich/analyze-list', {
-  method: 'POST',
-  body: JSON.stringify({ category, items, providerId, model })
-});
-export const saveEnrichmentList = (category, content, items) => request('/digital-twin/enrich/save-list', {
-  method: 'POST',
-  body: JSON.stringify({ category, content, items })
-});
-export const getEnrichmentListItems = (category) => request(`/digital-twin/enrich/list-items/${category}`);
-
-// --- Digital Twin Traits & Confidence (Phase 1 & 2) ---
-export const getDigitalTwinTraits = () => request('/digital-twin/traits');
-export const analyzeDigitalTwinTraits = (providerId, model, forceReanalyze = false) => request('/digital-twin/traits/analyze', {
-  method: 'POST',
-  body: JSON.stringify({ providerId, model, forceReanalyze })
-});
-export const updateDigitalTwinTraits = (updates) => request('/digital-twin/traits', {
-  method: 'PUT',
-  body: JSON.stringify(updates)
-});
-export const getDigitalTwinConfidence = () => request('/digital-twin/confidence');
-export const calculateDigitalTwinConfidence = (providerId, model) => request('/digital-twin/confidence/calculate', {
-  method: 'POST',
-  body: JSON.stringify({ providerId, model })
-});
-export const getDigitalTwinGaps = () => request('/digital-twin/gaps');
-
-// --- Digital Twin External Import (Phase 4) ---
-export const getDigitalTwinImportSources = () => request('/digital-twin/import/sources');
-export const analyzeDigitalTwinImport = (source, data, providerId, model) => request('/digital-twin/import/analyze', {
-  method: 'POST',
-  body: JSON.stringify({ source, data, providerId, model })
-});
-export const saveDigitalTwinImport = (source, suggestedDoc) => request('/digital-twin/import/save', {
-  method: 'POST',
-  body: JSON.stringify({ source, suggestedDoc })
-});
-
-// Digital Twin - Behavioral Feedback Loop
-export const submitBehavioralFeedback = (data) => request('/digital-twin/feedback', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const getBehavioralFeedbackStats = () => request('/digital-twin/feedback/stats');
-export const recalculateFeedbackWeights = () => request('/digital-twin/feedback/recalculate', {
-  method: 'POST'
-});
-export const getRecentFeedback = (contentType, limit) => {
-  const params = new URLSearchParams();
-  if (contentType) params.set('contentType', contentType);
-  if (limit) params.set('limit', limit);
-  return request(`/digital-twin/feedback/recent?${params}`);
-};
-
-// Digital Twin - Taste Questionnaire
-export const getTasteProfile = () => request('/digital-twin/taste');
-export const getTasteSections = () => request('/digital-twin/taste/sections');
-export const getTasteNextQuestion = (section) => request(`/digital-twin/taste/${section}/next`);
-export const submitTasteAnswer = (section, questionId, answer, meta = {}) => request('/digital-twin/taste/answer', {
-  method: 'POST',
-  body: JSON.stringify({ section, questionId, answer, ...meta })
-});
-export const getTasteSectionResponses = (section) => request(`/digital-twin/taste/${section}/responses`);
-export const generateTasteSummary = (providerId, model, section) => request('/digital-twin/taste/summary', {
-  method: 'POST',
-  body: JSON.stringify({ providerId, model, ...(section ? { section } : {}) })
-});
-export const getPersonalizedTasteQuestion = (section, providerId, model) =>
-  request(`/digital-twin/taste/${section}/personalized-question`, {
-    method: 'POST',
-    body: JSON.stringify({ providerId, model })
-  });
-export const resetTasteSection = (section) => request(`/digital-twin/taste/${section}`, {
-  method: 'DELETE'
-});
-
-// Digital Twin - Autobiography
-export const getAutobiographyStats = () => request('/digital-twin/autobiography');
-export const getAutobiographyConfig = () => request('/digital-twin/autobiography/config');
-export const updateAutobiographyConfig = (config) => request('/digital-twin/autobiography/config', {
-  method: 'PUT',
-  body: JSON.stringify(config)
-});
-export const getAutobiographyThemes = () => request('/digital-twin/autobiography/themes');
-export const getAutobiographyPrompt = (exclude) =>
-  request(`/digital-twin/autobiography/prompt${exclude ? `?exclude=${exclude}` : ''}`);
-export const getAutobiographyPromptById = (id) => request(`/digital-twin/autobiography/prompt/${id}`);
-export const getAutobiographyStories = (theme = null) =>
-  request(`/digital-twin/autobiography/stories${theme ? `?theme=${theme}` : ''}`);
-export const saveAutobiographyStory = (promptId, content, { parentStoryId, customPromptText } = {}) =>
-  request('/digital-twin/autobiography/stories', {
-    method: 'POST',
-    body: JSON.stringify({ promptId, content, parentStoryId, customPromptText })
-  });
-export const updateAutobiographyStory = (id, content) => request(`/digital-twin/autobiography/stories/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify({ content })
-});
-export const deleteAutobiographyStory = (id) => request(`/digital-twin/autobiography/stories/${id}`, {
-  method: 'DELETE'
-});
-export const triggerAutobiographyPrompt = () => request('/digital-twin/autobiography/trigger', {
-  method: 'POST'
-});
-export const generateAutobiographyFollowUps = (storyId, providerId) =>
-  request(`/digital-twin/autobiography/stories/${storyId}/follow-ups`, {
-    method: 'POST',
-    body: JSON.stringify({ providerId })
-  });
-export const getAutobiographyStoryChain = (storyId) =>
-  request(`/digital-twin/autobiography/stories/${storyId}/chain`);
-
-// Digital Twin - Assessment Analyzer
-export const analyzeAssessment = (content, providerId, model) =>
-  request('/digital-twin/interview/analyze', {
-    method: 'POST',
-    body: JSON.stringify({ content, providerId, model })
-  });
-
-// Digital Twin - Social Accounts
-export const getSocialAccounts = (params = {}) => {
-  const qs = new URLSearchParams(params).toString();
-  return request(`/digital-twin/social-accounts${qs ? `?${qs}` : ''}`);
-};
-export const getSocialAccountPlatforms = () => request('/digital-twin/social-accounts/platforms');
-export const getSocialAccountStats = () => request('/digital-twin/social-accounts/stats');
-export const getSocialAccount = (id) => request(`/digital-twin/social-accounts/${id}`);
-export const createSocialAccount = (data) => request('/digital-twin/social-accounts', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const createSocialAccountsBulk = (accounts) => request('/digital-twin/social-accounts/bulk', {
-  method: 'POST',
-  body: JSON.stringify({ accounts })
-});
-export const updateSocialAccount = (id, data) => request(`/digital-twin/social-accounts/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteSocialAccount = (id) => request(`/digital-twin/social-accounts/${id}`, {
-  method: 'DELETE'
-});
-
-// Digital Twin - Time Capsule Snapshots
-export const listTimeCapsuleSnapshots = () => request('/digital-twin/snapshots');
-export const createTimeCapsuleSnapshot = (label, description = '') => request('/digital-twin/snapshots', {
-  method: 'POST',
-  body: JSON.stringify({ label, description })
-});
-export const getTimeCapsuleSnapshot = (id) => request(`/digital-twin/snapshots/${id}`);
-export const deleteTimeCapsuleSnapshot = (id) => request(`/digital-twin/snapshots/${id}`, {
-  method: 'DELETE'
-});
-export const compareTimeCapsuleSnapshots = (id1, id2) => request('/digital-twin/snapshots/compare', {
-  method: 'POST',
-  body: JSON.stringify({ id1, id2 })
-});
-
-// MeatSpace - Genome
-export const getGenomeSummary = () => request('/meatspace/genome');
-export const uploadGenomeFile = (content, filename) => request('/meatspace/genome/upload', {
-  method: 'POST',
-  body: JSON.stringify({ content, filename })
-});
-export const scanGenomeMarkers = () => request('/meatspace/genome/scan', { method: 'POST' });
-export const searchGenomeSNP = (rsid) => request('/meatspace/genome/search', {
-  method: 'POST',
-  body: JSON.stringify({ rsid })
-});
-export const saveGenomeMarker = (data) => request('/meatspace/genome/markers', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateGenomeMarkerNotes = (id, notes) => request(`/meatspace/genome/markers/${id}/notes`, {
-  method: 'PUT',
-  body: JSON.stringify({ notes })
-});
-export const deleteGenomeMarker = (id) => request(`/meatspace/genome/markers/${id}`, { method: 'DELETE' });
-export const deleteGenomeData = () => request('/meatspace/genome', { method: 'DELETE' });
-
-// MeatSpace - Genome ClinVar
-export const getClinvarStatus = () => request('/meatspace/genome/clinvar/status');
-export const syncClinvar = () => request('/meatspace/genome/clinvar/sync', { method: 'POST' });
-export const scanClinvar = () => request('/meatspace/genome/clinvar/scan', { method: 'POST' });
-export const deleteClinvar = () => request('/meatspace/genome/clinvar', { method: 'DELETE' });
-
-// MeatSpace - Epigenetic Lifestyle Tracking
-export const getEpigeneticInterventions = () => request('/meatspace/genome/epigenetic');
-export const getEpigeneticRecommendations = (categories = []) =>
-  request(`/meatspace/genome/epigenetic/recommendations${categories.length ? `?categories=${categories.join(',')}` : ''}`);
-export const getEpigeneticCompliance = (days = 30) =>
-  request(`/meatspace/genome/epigenetic/compliance?days=${days}`);
-export const addEpigeneticIntervention = (data) => request('/meatspace/genome/epigenetic', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const logEpigeneticEntry = (id, entry) => request(`/meatspace/genome/epigenetic/${id}/log`, {
-  method: 'POST',
-  body: JSON.stringify(entry)
-});
-export const updateEpigeneticIntervention = (id, updates) => request(`/meatspace/genome/epigenetic/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(updates)
-});
-export const deleteEpigeneticIntervention = (id) => request(`/meatspace/genome/epigenetic/${id}`, {
-  method: 'DELETE'
-});
-
-// Digital Twin - Identity
-export const getIdentityStatus = () => request('/digital-twin/identity');
-export const getCrossInsights = () => request('/digital-twin/identity/cross-insights');
-export const getChronotype = () => request('/digital-twin/identity/chronotype');
-export const getChronotypeEnergySchedule = () => request('/digital-twin/identity/chronotype/energy-schedule');
-export const deriveChronotype = () => request('/digital-twin/identity/chronotype/derive', { method: 'POST' });
-export const updateChronotypeBehavioral = (data) => request('/digital-twin/identity/chronotype', {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const getLongevity = () => request('/digital-twin/identity/longevity');
-export const deriveLongevity = () => request('/digital-twin/identity/longevity/derive', { method: 'POST' });
-export const getGoals = (options) => request('/digital-twin/identity/goals', options);
-export const setBirthDate = (birthDate) => request('/digital-twin/identity/goals/birth-date', {
-  method: 'PUT',
-  body: JSON.stringify({ birthDate })
-});
-export const createGoal = (data) => request('/digital-twin/identity/goals', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateGoal = (id, data) => request(`/digital-twin/identity/goals/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteGoal = (id) => request(`/digital-twin/identity/goals/${id}`, { method: 'DELETE' });
-export const getGoalsTree = () => request('/digital-twin/identity/goals/tree');
-export const addGoalMilestone = (goalId, data) => request(`/digital-twin/identity/goals/${goalId}/milestones`, {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const completeGoalMilestone = (goalId, milestoneId) =>
-  request(`/digital-twin/identity/goals/${goalId}/milestones/${milestoneId}/complete`, { method: 'PUT' });
-export const linkGoalActivity = (goalId, data) => request(`/digital-twin/identity/goals/${goalId}/activities`, {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const unlinkGoalActivity = (goalId, activityName) =>
-  request(`/digital-twin/identity/goals/${goalId}/activities/${encodeURIComponent(activityName)}`, { method: 'DELETE' });
-export const addGoalProgress = (goalId, data) => request(`/digital-twin/identity/goals/${goalId}/progress`, {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const deleteGoalProgress = (goalId, entryId) =>
-  request(`/digital-twin/identity/goals/${goalId}/progress/${entryId}`, { method: 'DELETE' });
-
-// Feature Agents
-export const getFeatureAgents = () => request('/feature-agents');
-export const getFeatureAgent = (id) => request(`/feature-agents/${id}`);
-export const createFeatureAgent = (data) => request('/feature-agents', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateFeatureAgent = (id, data) => request(`/feature-agents/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteFeatureAgent = (id) => request(`/feature-agents/${id}`, { method: 'DELETE' });
-export const startFeatureAgent = (id) => request(`/feature-agents/${id}/start`, { method: 'POST' });
-export const pauseFeatureAgent = (id) => request(`/feature-agents/${id}/pause`, { method: 'POST' });
-export const resumeFeatureAgent = (id) => request(`/feature-agents/${id}/resume`, { method: 'POST' });
-export const triggerFeatureAgent = (id) => request(`/feature-agents/${id}/trigger`, { method: 'POST' });
-export const stopFeatureAgent = (id) => request(`/feature-agents/${id}/stop`, { method: 'POST' });
-export const getFeatureAgentRuns = (id, limit) => request(`/feature-agents/${id}/runs${limit ? `?limit=${limit}` : ''}`);
-export const getFeatureAgentOutput = (id) => request(`/feature-agents/${id}/output`);
-
-// MeatSpace - Health Tracker
-export const getMeatspaceOverview = () => request('/meatspace');
-export const getMeatspaceConfig = () => request('/meatspace/config');
-export const updateMeatspaceConfig = (data) => request('/meatspace/config', {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const updateMeatspaceLifestyle = (data) => request('/meatspace/lifestyle', {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const getMeatspaceBirthDate = () => request('/meatspace/birth-date');
-export const setMeatspaceBirthDate = (birthDate) => request('/meatspace/birth-date', {
-  method: 'PUT',
-  body: JSON.stringify({ birthDate })
-});
-export const getDeathClock = () => request('/meatspace/death-clock');
-export const getLEV = () => request('/meatspace/lev');
-export const getAlcoholSummary = () => request('/meatspace/alcohol');
-export const getDailyAlcohol = (from, to) => {
-  const params = new URLSearchParams();
-  if (from) params.set('from', from);
-  if (to) params.set('to', to);
-  return request(`/meatspace/alcohol/daily?${params}`);
-};
-export const logAlcoholDrink = (data) => request('/meatspace/alcohol/log', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateAlcoholDrink = (date, index, data) => request(`/meatspace/alcohol/log/${date}/${index}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const removeAlcoholDrink = (date, index) => request(`/meatspace/alcohol/log/${date}/${index}`, {
-  method: 'DELETE'
-});
-export const getCustomDrinks = () => request('/meatspace/alcohol/custom-drinks');
-export const addCustomDrink = (data) => request('/meatspace/alcohol/custom-drinks', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateCustomDrink = (index, data) => request(`/meatspace/alcohol/custom-drinks/${index}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const removeCustomDrink = (index) => request(`/meatspace/alcohol/custom-drinks/${index}`, {
-  method: 'DELETE'
-});
-export const getNicotineSummary = () => request('/meatspace/nicotine');
-export const getDailyNicotine = (from, to) => {
-  const params = new URLSearchParams();
-  if (from) params.set('from', from);
-  if (to) params.set('to', to);
-  return request(`/meatspace/nicotine/daily?${params}`);
-};
-export const logNicotine = (data) => request('/meatspace/nicotine/log', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateNicotineEntry = (date, index, data) => request(`/meatspace/nicotine/log/${date}/${index}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const removeNicotineEntry = (date, index) => request(`/meatspace/nicotine/log/${date}/${index}`, {
-  method: 'DELETE'
-});
-export const getCustomNicotineProducts = () => request('/meatspace/nicotine/custom-products');
-export const addCustomNicotineProduct = (data) => request('/meatspace/nicotine/custom-products', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateCustomNicotineProduct = (index, data) => request(`/meatspace/nicotine/custom-products/${index}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const removeCustomNicotineProduct = (index) => request(`/meatspace/nicotine/custom-products/${index}`, {
-  method: 'DELETE'
-});
-export const getBloodTests = () => request('/meatspace/blood');
-export const addBloodTest = (data) => request('/meatspace/blood', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const getBodyHistory = () => request('/meatspace/body');
-export const addBodyEntry = (data) => request('/meatspace/body', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const getEpigeneticTests = () => request('/meatspace/epigenetic');
-export const addEpigeneticTest = (data) => request('/meatspace/epigenetic', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const getEyeExams = () => request('/meatspace/eyes');
-export const addEyeExam = (data) => request('/meatspace/eyes', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateEyeExam = (index, data) => request(`/meatspace/eyes/${index}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const removeEyeExam = (index) => request(`/meatspace/eyes/${index}`, {
-  method: 'DELETE'
-});
-
-// MeatSpace - POST (Power On Self Test)
-export const getPostConfig = () => request('/meatspace/post/config');
-export const updatePostConfig = (data) => request('/meatspace/post/config', {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const getPostSessions = (from, to) => {
-  const params = new URLSearchParams();
-  if (from) params.set('from', from);
-  if (to) params.set('to', to);
-  return request(`/meatspace/post/sessions?${params}`);
-};
-export const getPostSession = (id) => request(`/meatspace/post/sessions/${id}`);
-export const submitPostSession = (data) => request('/meatspace/post/sessions', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const getPostStats = (days) => request(`/meatspace/post/stats${days != null ? `?days=${days}` : ''}`);
-export const generatePostDrill = (type, config = {}, providerId, model) => request('/meatspace/post/drill', {
-  method: 'POST',
-  body: JSON.stringify({ type, config, ...(providerId && { providerId }), ...(model && { model }) })
-});
-export const scorePostLlmDrill = (type, drillData, responses, timeLimitMs, providerId, model) =>
-  request('/meatspace/post/score-llm', {
-    method: 'POST',
-    body: JSON.stringify({ type, drillData, responses, timeLimitMs, ...(providerId && { providerId }), ...(model && { model }) })
-  });
-
-// MeatSpace - POST Memory Builder
-export const getMemoryItems = () => request('/meatspace/post/memory-items');
-export const getMemoryItem = (id) => request(`/meatspace/post/memory-items/${id}`);
-export const createMemoryItem = (data) => request('/meatspace/post/memory-items', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateMemoryItem = (id, data) => request(`/meatspace/post/memory-items/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteMemoryItem = (id) => request(`/meatspace/post/memory-items/${id}`, {
-  method: 'DELETE'
-});
-export const submitMemoryPractice = (id, data) => request(`/meatspace/post/memory-items/${id}/practice`, {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const getMemoryMastery = (id) => request(`/meatspace/post/memory-items/${id}/mastery`);
-export const getChunkMastery = (id) => request(`/meatspace/post/memory-items/${id}/chunk-mastery`);
-export const generateMemoryDrill = (data) => request('/meatspace/post/memory-drill', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-
-// MeatSpace - POST Training Log
-export const submitTrainingEntry = (data) => request('/meatspace/post/training', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const getTrainingStats = (days) => request(`/meatspace/post/training/stats${days != null ? `?days=${days}` : ''}`);
-export const getTrainingEntries = (limit) => request(`/meatspace/post/training/entries${limit ? `?limit=${limit}` : ''}`);
-
-// Life Calendar
-export const getLifeCalendar = () => request('/meatspace/calendar');
-export const getActivities = () => request('/meatspace/activities');
-export const addActivity = (data) => request('/meatspace/activities', {
-  method: 'POST', body: JSON.stringify(data)
-});
-export const updateActivity = (index, data) => request(`/meatspace/activities/${index}`, {
-  method: 'PUT', body: JSON.stringify(data)
-});
-export const removeActivity = (index) => request(`/meatspace/activities/${index}`, { method: 'DELETE' });
-
-// Life Events
-export const getLifeEvents = () => request('/meatspace/life-events');
-export const addLifeEvent = (data) => request('/meatspace/life-events', {
-  method: 'POST', body: JSON.stringify(data)
-});
-export const updateLifeEvent = (id, data) => request(`/meatspace/life-events/${id}`, {
-  method: 'PUT', body: JSON.stringify(data)
-});
-export const removeLifeEvent = (id) => request(`/meatspace/life-events/${id}`, { method: 'DELETE' });
-
-// DataDog
-export const getDatadogInstances = () => request('/datadog/instances');
-
-// JIRA
-export const getJiraInstances = () => request('/jira/instances');
-export const getJiraProjects = (instanceId) => request(`/jira/instances/${instanceId}/projects`);
-export const getMySprintTickets = (instanceId, projectKey) => request(`/jira/instances/${instanceId}/my-sprint-tickets/${projectKey}`);
-export const getJiraTicketTransitions = (instanceId, ticketId, options) => request(`/jira/instances/${instanceId}/tickets/${ticketId}/transitions`, options);
-export const transitionJiraTicket = (instanceId, ticketId, transitionId, options) => request(`/jira/instances/${instanceId}/tickets/${ticketId}/transition`, {
-  method: 'POST',
-  body: JSON.stringify({ transitionId }),
-  ...options
-});
-
-// JIRA Status Reports
-export const getJiraReports = () => request('/jira/reports');
-export const generateJiraReport = (appId) => request('/jira/reports/generate', {
-  method: 'POST',
-  body: JSON.stringify(appId ? { appId } : {})
-});
-export const getJiraReport = (appId, date) => request(`/jira/reports/${appId}/${date}`);
-export const getLatestJiraReport = (appId) => request(`/jira/reports/${appId}/latest`);
-
-// Browser - CDP browser management
-export const getBrowserStatus = () => request('/browser');
-export const getBrowserConfig = () => request('/browser/config');
-export const updateBrowserConfig = (config) => request('/browser/config', {
-  method: 'PUT',
-  body: JSON.stringify(config)
-});
-export const launchBrowser = () => request('/browser/launch', { method: 'POST' });
-export const stopBrowser = () => request('/browser/stop', { method: 'POST' });
-export const restartBrowser = () => request('/browser/restart', { method: 'POST' });
-export const getBrowserHealth = () => request('/browser/health');
-export const getBrowserProcess = () => request('/browser/process');
-export const getBrowserPages = () => request('/browser/pages');
-export const getBrowserVersion = () => request('/browser/version');
-export const getBrowserLogs = (lines = 50) => request(`/browser/logs?lines=${lines}`);
-export const getBrowserDownloads = () => request('/browser/downloads');
-export const navigateBrowser = (url) => request('/browser/navigate', {
-  method: 'POST',
-  body: JSON.stringify({ url })
-});
-
-// Backup
-export const getBackupStatus = (options) => request('/backup/status', options);
-export const triggerBackup = () => request('/backup/run', { method: 'POST' });
-export const getBackupSnapshots = (options) => request('/backup/snapshots', options);
-export const restoreBackup = (data) => request('/backup/restore', { method: 'POST', body: JSON.stringify(data) });
-
-// Settings
-export const getSettings = () => request('/settings');
-export const updateSettings = (data) => request('/settings', {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-
-// Telegram
-export const getTelegramStatus = () => request('/telegram/status');
-export const updateTelegramConfig = (data) => request('/telegram/config', {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteTelegramConfig = () => request('/telegram/config', { method: 'DELETE' });
-export const testTelegram = (message) => request('/telegram/test', {
-  method: 'POST',
-  body: JSON.stringify({ message })
-});
-export const updateTelegramForwardTypes = (forwardTypes) => request('/telegram/forward-types', {
-  method: 'PUT',
-  body: JSON.stringify({ forwardTypes })
-});
-export const updateTelegramMethod = (method) => request('/telegram/method', {
-  method: 'PUT',
-  body: JSON.stringify({ method })
-});
-export const reloadTelegramBridge = () => request('/telegram/bridge/reload', { method: 'POST' });
-
-// Image Generation
-export const getImageGenStatus = () => request('/image-gen/status');
-export const generateImage = (data) => request('/image-gen/generate', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const generateAvatar = (data) => request('/image-gen/avatar', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-
-// Tools Registry
-export const getToolsList = () => request('/tools');
-export const getEnabledTools = () => request('/tools/enabled');
-export const registerTool = (data) => request('/tools', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateTool = (id, data) => request(`/tools/${id}`, {
-  method: 'PUT',
-  body: JSON.stringify(data)
-});
-export const deleteTool = (id) => request(`/tools/${id}`, { method: 'DELETE' });
-
-// Insights
-export const getGenomeHealthCorrelations = () => request('/insights/genome-health');
-export const getInsightThemes = () => request('/insights/themes');
-export const refreshInsightThemes = (providerId, model) => request('/insights/themes/refresh', {
-  method: 'POST',
-  body: JSON.stringify({ providerId, model })
-});
-export const getInsightNarrative = () => request('/insights/narrative');
-export const refreshInsightNarrative = (providerId, model) => request('/insights/narrative/refresh', {
-  method: 'POST',
-  body: JSON.stringify({ providerId, model })
-});
-
-// Instances (Federation)
-export const getInstances = () => request('/instances');
-export const getSelfInstance = () => request('/instances/self');
-export const updateSelfInstance = (data) => request('/instances/self', { method: 'PUT', body: JSON.stringify(data) });
-export const addPeer = (data) => request('/instances/peers', { method: 'POST', body: JSON.stringify(data) });
-export const updatePeer = (id, data) => request(`/instances/peers/${id}`, { method: 'PUT', body: JSON.stringify(data) });
-export const removePeer = (id) => request(`/instances/peers/${id}`, { method: 'DELETE' });
-export const connectPeer = (id) => request(`/instances/peers/${id}/connect`, { method: 'POST' });
-export const probePeer = (id) => request(`/instances/peers/${id}/probe`, { method: 'POST' });
-export const queryPeer = (id, path) => request(`/instances/peers/${id}/query?path=${encodeURIComponent(path)}`);
-
-// Data Manager
-export const getDataOverview = () => request('/data');
-export const getDataCategory = (key) => request(`/data/${key}`);
-export const archiveDataCategory = (key, opts) => request(`/data/${key}/archive`, { method: 'POST', body: JSON.stringify(opts || {}) });
-export const purgeDataCategory = (key, opts) => request(`/data/${key}`, { method: 'DELETE', body: JSON.stringify(opts || {}) });
-export const getDataBackups = () => request('/data/backups');
-export const deleteDataBackup = (filename) => request(`/data/backups/${filename}`, { method: 'DELETE' });
-
-// GSD (Get Stuff Done) Integration
-export const getGsdProjects = () => request('/cos/gsd/projects');
-export const getGsdProject = (appId) => request(`/cos/gsd/projects/${appId}`);
-export const getGsdConcerns = (appId) => request(`/cos/gsd/projects/${appId}/concerns`);
-export const getGsdPhases = (appId) => request(`/cos/gsd/projects/${appId}/phases`);
-export const getGsdPhase = (appId, phaseId) => request(`/cos/gsd/projects/${appId}/phases/${phaseId}`);
-export const createGsdConcernTasks = (appId, data) => request(`/cos/gsd/projects/${appId}/concerns/tasks`, {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const triggerGsdPhaseAction = (appId, phaseId, action) => request(`/cos/gsd/projects/${appId}/phases/${phaseId}/action`, {
-  method: 'POST',
-  body: JSON.stringify({ action })
-});
-export const getGsdDocument = (appId, docName) => request(`/cos/gsd/projects/${appId}/documents/${docName}`);
-export const saveGsdDocument = (appId, docName, content, commitMessage) => request(`/cos/gsd/projects/${appId}/documents/${docName}`, {
-  method: 'PUT',
-  body: JSON.stringify({ content, ...(commitMessage && { commitMessage }) })
-});
-
-// GitHub Repos
-export const getGitHubRepos = () => request('/github/repos');
-export const syncGitHubRepos = () => request('/github/repos/sync', { method: 'POST' });
-export const updateGitHubRepo = (fullName, data) =>
-  request(`/github/repos/${encodeURIComponent(fullName)}`, { method: 'PUT', body: JSON.stringify(data) });
-export const archiveGitHubRepo = (fullName) =>
-  request(`/github/repos/${encodeURIComponent(fullName)}/archive`, { method: 'POST' });
-export const unarchiveGitHubRepo = (fullName) =>
-  request(`/github/repos/${encodeURIComponent(fullName)}/unarchive`, { method: 'POST' });
-export const getGitHubSecrets = () => request('/github/secrets');
-export const setGitHubSecret = (name, value) =>
-  request(`/github/secrets/${encodeURIComponent(name)}`, { method: 'PUT', body: JSON.stringify({ value }) });
-export const syncGitHubSecret = (name) =>
-  request(`/github/secrets/${encodeURIComponent(name)}/sync`, { method: 'POST' });
-
-// Messages
-export const getMessageAccounts = () => request('/messages/accounts');
-export const createMessageAccount = (data) => request('/messages/accounts', { method: 'POST', body: JSON.stringify(data) });
-export const updateMessageAccount = (id, data) => request(`/messages/accounts/${id}`, { method: 'PUT', body: JSON.stringify(data) });
-export const deleteMessageAccount = (id) => request(`/messages/accounts/${id}`, { method: 'DELETE' });
-export const syncMessageAccount = (accountId, mode = 'unread') => request(`/messages/sync/${accountId}`, { method: 'POST', body: JSON.stringify({ mode }) });
-export const getMessageSyncStatus = (accountId) => request(`/messages/sync/${accountId}/status`);
-export const evaluateMessages = (data = {}) => request('/messages/evaluate', { method: 'POST', body: JSON.stringify(data) });
-export const getMessageInbox = (params = {}) => {
-  const qs = new URLSearchParams();
-  if (params.accountId) qs.set('accountId', params.accountId);
-  if (params.search) qs.set('search', params.search);
-  if (params.limit) qs.set('limit', params.limit);
-  if (params.offset) qs.set('offset', params.offset);
-  const str = qs.toString();
-  return request(`/messages/inbox${str ? `?${str}` : ''}`);
-};
-export const getMessageDetail = (accountId, messageId) => request(`/messages/${accountId}/${messageId}`);
-export const getMessageThread = (accountId, threadId) => request(`/messages/thread/${accountId}/${threadId}`);
-export const getMessageDrafts = (params = {}) => {
-  const qs = new URLSearchParams();
-  if (params.accountId) qs.set('accountId', params.accountId);
-  if (params.status) qs.set('status', params.status);
-  const str = qs.toString();
-  return request(`/messages/drafts${str ? `?${str}` : ''}`);
-};
-export const createMessageDraft = (data) => request('/messages/drafts', { method: 'POST', body: JSON.stringify(data) });
-export const generateMessageDraft = (data) => request('/messages/drafts/generate', { method: 'POST', body: JSON.stringify(data) });
-export const updateMessageDraft = (id, data) => request(`/messages/drafts/${id}`, { method: 'PUT', body: JSON.stringify(data) });
-export const approveMessageDraft = (id) => request(`/messages/drafts/${id}/approve`, { method: 'POST' });
-export const sendMessageDraft = (id) => request(`/messages/drafts/${id}/send`, { method: 'POST' });
-export const deleteMessageDraft = (id) => request(`/messages/drafts/${id}`, { method: 'DELETE' });
-export const getMessageSelectors = () => request('/messages/selectors');
-export const updateMessageSelectors = (provider, selectors) => request(`/messages/selectors/${provider}`, { method: 'PUT', body: JSON.stringify({ selectors }) });
-export const testMessageSelectors = (provider) => request(`/messages/selectors/${provider}/test`, { method: 'POST' });
-export const launchMessageBrowser = (accountId) => request(`/messages/launch/${accountId}`, { method: 'POST' });
-export const refreshMessage = (accountId, messageId) =>
-  request(`/messages/${accountId}/${messageId}/refresh`, { method: 'POST' });
-export const fetchFullContent = (accountId, { force } = {}) =>
-  request(`/messages/fetch-full/${accountId}`, { method: 'POST', body: force ? JSON.stringify({ force: true }) : undefined });
-export const executeMessageAction = (accountId, messageId, action) =>
-  request(`/messages/${accountId}/${messageId}/action`, { method: 'POST', body: JSON.stringify({ action }), silent: true });
-export const clearMessageCache = (accountId) =>
-  request(`/messages/accounts/${accountId}/cache/clear`, { method: 'POST' });
-export const enableGmailApi = () => request('/messages/gmail/enable-api', { method: 'POST' });
-
-// Calendar
-export const getCalendarAccounts = () => request('/calendar/accounts');
-export const createCalendarAccount = (data) => request('/calendar/accounts', { method: 'POST', body: JSON.stringify(data) });
-export const updateCalendarAccount = (id, data) => request(`/calendar/accounts/${id}`, { method: 'PUT', body: JSON.stringify(data) });
-export const deleteCalendarAccount = (id) => request(`/calendar/accounts/${id}`, { method: 'DELETE' });
-export const syncCalendarAccount = (accountId) => request(`/calendar/sync/${accountId}`, { method: 'POST' });
-export const getCalendarSyncStatus = (accountId) => request(`/calendar/sync/${accountId}/status`);
-export const getCalendarEvents = (params = {}) => {
-  const str = new URLSearchParams(Object.entries(params).filter(([, v]) => v != null)).toString();
-  return request(`/calendar/events${str ? `?${str}` : ''}`);
-};
-export const getCalendarTokenStatus = () => request('/calendar/debug/token-status');
-export const testCalendarToken = (provider) => request('/calendar/debug/test-token', { method: 'POST', body: JSON.stringify({ provider }) });
-export const clearCalendarToken = (provider) => request('/calendar/debug/clear-token', { method: 'POST', body: JSON.stringify({ provider }) });
-export const updateSubcalendars = (accountId, data) => request(`/calendar/accounts/${accountId}/subcalendars`, { method: 'PUT', body: JSON.stringify(data) });
-export const mcpSyncGoogleCalendar = (accountId) => request(`/calendar/sync/${accountId}/google`, { method: 'POST' });
-export const mcpDiscoverCalendars = (accountId) => request(`/calendar/sync/${accountId}/discover`, { method: 'POST' });
-export const getGoogleAuthStatus = () => request('/calendar/google/auth/status');
-export const saveGoogleAuthCredentials = (data) => request('/calendar/google/auth/credentials', { method: 'POST', body: JSON.stringify(data) });
-export const getGoogleAuthUrl = () => request('/calendar/google/auth/url');
-export const clearGoogleAuth = () => request('/calendar/google/auth/clear', { method: 'POST' });
-export const apiSyncGoogleCalendar = (accountId) => request(`/calendar/sync/${accountId}/api`, { method: 'POST' });
-export const apiDiscoverCalendars = (accountId) => request(`/calendar/sync/${accountId}/discover-api`, { method: 'POST' });
-export const startGoogleAutoConfig = () => request('/calendar/google/auto-configure/start', { method: 'POST' });
-export const runGoogleAutoConfig = (email) => request('/calendar/google/auto-configure/run', { method: 'POST', body: JSON.stringify({ email }) });
-export const getDailyReview = (date) => request(`/calendar/review/${date}`);
-export const confirmDailyReviewEvent = (date, data) => request(`/calendar/review/${date}/confirm`, { method: 'POST', body: JSON.stringify(data) });
-export const getDailyReviewHistory = (params = {}) => {
-  const str = new URLSearchParams(Object.entries(params).filter(([, v]) => v)).toString();
-  return request(`/calendar/review/history${str ? `?${str}` : ''}`);
-};
-export const linkGoalCalendar = (goalId, data) => request(`/digital-twin/identity/goals/${goalId}/calendars`, { method: 'POST', body: JSON.stringify(data) });
-export const unlinkGoalCalendar = (goalId, subcalendarId) => request(`/digital-twin/identity/goals/${goalId}/calendars/${encodeURIComponent(subcalendarId)}`, { method: 'DELETE' });
-export const getGoalCalendarEvents = (goalId, params = {}) => {
-  const str = new URLSearchParams(Object.entries(params).filter(([, v]) => v)).toString();
-  return request(`/digital-twin/identity/goals/${goalId}/calendar-events${str ? `?${str}` : ''}`);
-};
-
-// Goal Progress & Todos
-export const updateGoalProgress = (goalId, value) => request(`/digital-twin/identity/goals/${goalId}/progress`, { method: 'PUT', body: JSON.stringify({ value }) });
-export const addGoalTodo = (goalId, data) => request(`/digital-twin/identity/goals/${goalId}/todos`, { method: 'POST', body: JSON.stringify(data) });
-export const updateGoalTodo = (goalId, todoId, data) => request(`/digital-twin/identity/goals/${goalId}/todos/${todoId}`, { method: 'PUT', body: JSON.stringify(data) });
-export const deleteGoalTodo = (goalId, todoId) => request(`/digital-twin/identity/goals/${goalId}/todos/${todoId}`, { method: 'DELETE' });
-
-// Goal AI Planning & Scheduling
-export const generateGoalPhases = (goalId, options = {}) => request(`/digital-twin/identity/goals/${goalId}/generate-phases`, { method: 'POST', body: JSON.stringify(options) });
-export const acceptGoalPhases = (goalId, phases) => request(`/digital-twin/identity/goals/${goalId}/accept-phases`, { method: 'POST', body: JSON.stringify({ phases }) });
-export const organizeGoals = (options = {}) => request('/digital-twin/identity/goals/organize', { method: 'POST', body: JSON.stringify(options) });
-export const applyGoalOrganization = (organization) => request('/digital-twin/identity/goals/organize/apply', { method: 'POST', body: JSON.stringify({ organization }) });
-export const checkInGoal = (goalId, options = {}) => request(`/digital-twin/identity/goals/${goalId}/check-in`, { method: 'POST', body: JSON.stringify(options) });
-export const scheduleGoalTimeBlocks = (goalId) => request(`/digital-twin/identity/goals/${goalId}/schedule`, { method: 'POST' });
-export const removeGoalSchedule = (goalId) => request(`/digital-twin/identity/goals/${goalId}/schedule`, { method: 'DELETE' });
-export const rescheduleGoalTimeBlocks = (goalId) => request(`/digital-twin/identity/goals/${goalId}/reschedule`, { method: 'POST' });
-
-// Review Hub
-export const getReviewItems = (params) => {
-  const qs = new URLSearchParams();
-  if (params?.status) qs.set('status', params.status);
-  if (params?.type) qs.set('type', params.type);
-  const query = qs.toString();
-  return request(`/review/items${query ? `?${query}` : ''}`);
-};
-export const getReviewCounts = () => request('/review/counts');
-export const getReviewBriefing = () => request('/review/briefing');
-export const createReviewTodo = (data) => request('/review/todo', {
-  method: 'POST',
-  body: JSON.stringify(data)
-});
-export const updateReviewItem = (id, data) => request(`/review/items/${id}`, {
-  method: 'PATCH',
-  body: JSON.stringify(data)
-});
-export const completeReviewItem = (id) => request(`/review/items/${id}/complete`, { method: 'POST' });
-export const dismissReviewItem = (id) => request(`/review/items/${id}/dismiss`, { method: 'POST' });
-export const deleteReviewItem = (id) => request(`/review/items/${id}`, { method: 'DELETE' });
-
-// Default export for simplified imports
-export default {
-  get: (endpoint, options) => request(endpoint, { method: 'GET', ...options }),
-  post: (endpoint, body, options) => request(endpoint, {
-    method: 'POST',
-    body: JSON.stringify(body),
-    ...options
-  }),
-  put: (endpoint, body, options) => request(endpoint, {
-    method: 'PUT',
-    body: JSON.stringify(body),
-    ...options
-  }),
-  delete: (endpoint, options) => request(endpoint, { method: 'DELETE', ...options })
-};
-
-// Database
-export const getDatabaseStatus = () => request('/database/status');
-export const switchDatabase = (target, migrate = false) => request('/database/switch', {
-  method: 'POST',
-  body: JSON.stringify({ target, migrate })
-});
-export const setupNativeDatabase = () => request('/database/setup-native', { method: 'POST' });
-export const exportDatabase = (backend) => request('/database/export', {
-  method: 'POST',
-  ...(backend ? { body: JSON.stringify({ backend }) } : {})
-});
-export const fixDatabase = () => request('/database/fix', { method: 'POST' });
-export const syncDatabase = () => request('/database/sync', { method: 'POST' });
-export const startDatabase = (backend) => request('/database/start', {
-  method: 'POST',
-  body: JSON.stringify({ backend })
-});
-export const stopDatabase = (backend) => request('/database/stop', {
-  method: 'POST',
-  body: JSON.stringify({ backend })
-});
-export const destroyDatabase = (backend) => request('/database/destroy', {
-  method: 'POST',
-  body: JSON.stringify({ backend })
-});
-
-// Loops
-export const getLoops = () => request('/loops');
-export const getLoop = (id) => request(`/loops/${id}`);
-export const getLoopProviders = () => request('/loops/providers');
-export const createLoop = (data) => request('/loops', { method: 'POST', body: JSON.stringify(data) });
-export const updateLoop = (id, data) => request(`/loops/${id}`, { method: 'PUT', body: JSON.stringify(data) });
-export const stopLoop = (id) => request(`/loops/${id}/stop`, { method: 'POST' });
-export const resumeLoop = (id) => request(`/loops/${id}/resume`, { method: 'POST' });
-export const triggerLoop = (id) => request(`/loops/${id}/trigger`, { method: 'POST' });
-export const deleteLoop = (id) => request(`/loops/${id}`, { method: 'DELETE' });
+// Barrel file — re-exports all API sub-modules.
+// Consumers using `import * as api from '../services/api'` continue to work unchanged.
+
+export * from './apiCore.js';
+export * from './apiHealth.js';
+export * from './apiSystem.js';
+export * from './apiApps.js';
+export * from './apiPorts.js';
+export * from './apiScaffold.js';
+export * from './apiProviders.js';
+export * from './apiRuns.js';
+export * from './apiHistory.js';
+export * from './apiCommands.js';
+export * from './apiGit.js';
+export * from './apiMedia.js';
+export * from './apiAgents.js';
+export * from './apiPersonalities.js';
+export * from './apiAccounts.js';
+export * from './apiSchedules.js';
+export * from './apiMessages.js';
+export * from './apiCalendar.js';
+export * from './apiGoals.js';
+export * from './apiReview.js';
+export * from './apiDatabase.js';
+export * from './apiGSD.js';
+export * from './apiGithub.js';
+export * from './apiLoops.js';
+export * from './apiMemory.js';
+export * from './apiBrain.js';
+export * from './apiDigitalTwin.js';
+export * from './apiMeatspace.js';
+
+// Default export for simplified imports (get/post/put/delete helpers)
+export { default } from './apiCore.js';
