+++ package.json
@@ -18,7 +18,7 @@
     "pm2:status": "node ./node_modules/pm2/bin/pm2 status",
     "pm2:save": "node ./node_modules/pm2/bin/pm2 save",
     "pm2:startup": "node ./node_modules/pm2/bin/pm2 startup",
-    "install:all": "git submodule update --init --recursive && npm install && npm install --prefix client && npm install --prefix server && npm install --prefix autofixer && npm run setup",
+    "install:all": "git submodule update --init --recursive && npm install && npm install --prefix client && npm install --prefix server && npm install --prefix autofixer && node client/node_modules/esbuild/install.js && node server/node_modules/esbuild/install.js && npm rebuild node-pty --prefix server && npm run setup",
     "migrations": "node scripts/run-migrations.js",
     "setup": "node scripts/setup-data.js && node scripts/setup-db.js && node scripts/setup-browser.js",
     "setup:data": "node scripts/setup-data.js",
